+++ packages/cli/src/commands/add/index.ts
@@ -9,7 +9,7 @@ import type { Config } from "@changesets/types";
 import writeChangeset from "@changesets/write";
 import { ExitError } from "@changesets/errors";
 import { getPackages } from "@manypkg/get-packages";
-import { ExternalEditor } from "external-editor";
+import { ExternalEditor } from "@inquirer/external-editor";
 import { getCommitFunctions } from "../../commit/getCommitFunctions.ts";
 import * as cli from "../../utils/cli-utilities.ts";
 import { getVersionableChangedPackages } from "../../utils/versionablePackages.ts";